feat(callbacks): timeout on clear
In emulation, sometimes there is a deadlock in shutdown
1. rv_tester_terminate DPI export called
* The callback thread is hung waiting on this
2. rv_tester_shutdown_registry DPI import called
* This hangs waiting to join the callback thread

There seems to be another mutex in zebu that prevents #1 from finishing while in #2

This commit adds a timeout to the callback join. This allows whatever's calling shutdown to do so in a loop until successful, allowing the deadlock to be broken.