feat(packet_gen): add a new 'qualify' key to field
If the qualification is not met, that part of the packet is not sent.
This results in some overhead because we need to send 1-bit per unique qualification
But if some packets regularly have invalid fields, this can save a lot of DPI bandwidth
In rv_tester, cosim_m_rvfi was sending 162 bytes everytime. By qualifying certain fields the common packet rvfi size is 71 instead.